Fix asset bundle issue and made a class of them
