+++ NewHorizons/Builder/StarSystem/SkyboxBuilder.cs
@@ -10,44 +10,10 @@ namespace NewHorizons.Builder.StarSystem
 {
     public class SkyboxBuilder
     {
-        public static Material LoadMaterial(string assetBundle, string path, string uniqueModName, IModBehaviour mod)
-        {
-            string key = uniqueModName + "." + assetBundle;
-            AssetBundle bundle;
-            Material cubemap;
-
-            try
-            {
-                if (Main.AssetBundles.ContainsKey(key)) bundle = Main.AssetBundles[key];
-                else
-                {
-                    bundle = mod.ModHelper.Assets.LoadBundle(assetBundle);
-                    Main.AssetBundles[key] = bundle;
-                }
-            }
-            catch (Exception e)
-            {
-                Logger.LogError($"Couldn't load AssetBundle {assetBundle} : {e.Message}");
-                return null;
-            }
-
-            try
-            {
-                cubemap = bundle.LoadAsset<Material>(path);
-            }
-            catch (Exception e)
-            {
-                Logger.Log($"Couldn't load asset {path} from AssetBundle {assetBundle} : {e.Message}");
-                return null;
-            }
-
-            return cubemap;
-        }
-
         public static void Make(StarSystemConfig.SkyboxConfig info, IModBehaviour mod)
         {
             Logger.Log("Building Skybox");
-            Material skyBoxMaterial = LoadMaterial(info.assetBundle, info.path, mod.ModHelper.Manifest.UniqueName, mod);
+            var skyBoxMaterial = AssetBundleUtilities.Load<Material>(info.assetBundle, info.path, mod);
             RenderSettings.skybox = skyBoxMaterial;
             DynamicGI.UpdateEnvironment();
             Main.Instance.ModHelper.Events.Unity.FireOnNextUpdate(() =>
